:sparkles: works ??

+++ deploy/includes/validation.php
@@ -72,51 +72,43 @@ function validateFreetext($text):bool{
     $maxSize = 1024*1024; // = 1MB
     $allowedTypes = ['image/jpeg', 'image/png'];
 
-
     //PHPs own methods read the "magic information" off the file directly
-
     $fileInfo = finfo_open(FILEINFO_MIME_TYPE);
     $fileType = finfo_file($fileInfo, $file['tmp_name']);
     finfo_close($fileInfo);
 
-
     if (!in_array($fileType, $allowedTypes)) {
-
         //does not display the incorrect filetype to avoid educating attackers
-
         throw new UnexpectedValueException("Invalid file type.");
-
     }
-
     if ($file['size'] > $maxSize) {
-
         throw new LengthException("File exceeds maximum size.");
-
     }
-
     return $file['tmp_name'];
 } 
 
-/**
- * @param string $temporaryLocation Location of temporary file, for instance returned by "validateImage".
- * @param string $fileDestinationFolder Absolute filepath to the folder the file should be stored in
- * @param string $fileName 
- * moveUploadedFile($file, "/myComputer/folder/imagename.jpg")
+ /**
+ * @param string $temporaryLocation Path to the temporary file, returned by validateImage().
+ * @param string $fileDestinationFolder Target directory for storage.
+ * @throws RuntimeException If the file cannot be moved to target destination.
  */
-function moveUploadedFile(string $temporaryLocation, string $fileDestinationFolder, string $fileName):void{
-    $fileNameWithoutExtension = pathinfo($fileName, PATHINFO_FILENAME);
-    $fileType = finfo_file(finfo_open(FILEINFO_MIME_TYPE), $temporaryLocation);
+function moveUploadedFile(string $temporaryLocation, string $fileDestinationFolder):void{
 
+    $fileType = finfo_file(finfo_open(FILEINFO_MIME_TYPE), $temporaryLocation);
     $extensions = [
         "image/jpeg" => "jpg",
         "image/png" => "png"
     ];
-
     $extension = $extensions[$fileType] ?? "bin";
-    $fullpath = rtrim($fileDestinationFolder, "/") . "/" . $fileNameWithoutExtension . "." . $extension;   
 
+     //prevets duplicates and sanitizes - hash based on the file's binary data
+    $uniqueName = md5_file($temporaryLocation);
+    $fullpath = rtrim($fileDestinationFolder, "/") . "/" . $uniqueName . "." . $extension;  
+
+    if (file_exists($fullpath)) return;
     if (!move_uploaded_file($temporaryLocation, $fullpath)){
-            throw new RunTimeException("Failed to move $temporaryLocation to $fileDestinationFolder.");
+            throw new RuntimeException("Failed to move $temporaryLocation to $fileDestinationFolder.");
     }
-}
+
+}
